Consolidated index into Server
