+++ index.js
@@ -1,86 +1,2 @@
-import express from "express";
-import cors from "cors";
-import morgan from "morgan";
-import dotenv from "dotenv";
-import swaggerUI from "swagger-ui-express";
-import swaggerJsDoc from "swagger-jsdoc";
-
-import bodyParser from "body-parser";
-
-// Import the router from the hello.js file
-import postRouter from "./src/Routes/posts.js";
-import helloRouter from "./src/hello.js";
-
-// CDN CSS
-
-const CSS_URL =
-  "https://cdnjs.cloudflare.com/ajax/libs/swagger-ui/4.1.0/swagger-ui.min.css";
-
-const app = express();
-
-app.use(bodyParser.json()); // to use body object in requests
-const PORT = process.env.PORT || 8080;
-dotenv.config();
-
-app.use(morgan("dev"));
-app.use(cors());
-
-const options = {
-  definition: {
-    openapi: "3.0.0",
-    info: {
-      title: "Library API",
-      version: "1.0.0",
-      description: "A simple Express Library API",
-      termsOfService: "http://example.com/terms/",
-      contact: {
-        name: "API Support",
-        url: "http://www.exmaple.com/support",
-        email: "support@example.com",
-      },
-    },
-  },
-  // This is to call all the file
-  apis: ["src/**/*.js"],
-};
-
-const specs = swaggerJsDoc(options);
-// app.use("/api-docs", swaggerUI.serve, swaggerUI.setup(specs));
-
-app.use(
-  "/api-docs",
-  swaggerUI.serve,
-  swaggerUI.setup(specs, { customCssUrl: CSS_URL })
-);
-
-// Here we are calling the basic html
-// Use the router from the hello.js file
-app.use("/", helloRouter);
-// Use the router from the post.js file
-app.use("/posts", postRouter);
-
-// Endpoint to get available routes and methods
-app.get('/available-routes', (req, res) => {
-  const routes = [];
-  app._router.stack.forEach(middleware => {
-    if (middleware.route) { // routes registered directly on the app
-      routes.push({
-        path: middleware.route.path,
-        methods: Object.keys(middleware.route.methods).join(', ').toUpperCase()
-      });
-    } else if (middleware.name === 'router') { // router middleware
-      middleware.handle.stack.forEach(handler => {
-        const route = handler.route;
-        route && routes.push({
-          path: route.path,
-          methods: Object.keys(route.methods).join(', ').toUpperCase()
-        });
-      });
-    }
-  });
-  res.json(routes);
-});
-
-app.listen(PORT, () => console.log(`Server runs on port ${PORT}`));
-
+import server from "./src/server.js";
